chore(config): update port datatype setting

+++ infrastructure/meilisearch/config.go
@@ -2,7 +2,7 @@ package meilisearch
 
 type Config struct {
 	Host              string `yaml:"host"`
-	Port              int16  `yaml:"port"`
+	Port              int    `yaml:"port"`
 	Timeout           int16  `yaml:"timeout_in_ms"`
 	DefaultCollection string `yaml:"default_collection"`
 }
